pyproject.toml
[project]
name = "orbitmeetai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-docx",
    "ollama>=0.6.1",
    "pdfplumber>=0.10.3",
    "docx>=0.2.4",
    "langchain-google-genai>=3.2.0",
    "python-dotenv>=1.2.1",
    "langchain>=1.1.0",
    "langchain-groq>=1.1.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "pymongo>=4.15.4",
    "rapidfuzz>=3.14.3",
    "docx2txt>=0.9",
    "loguru>=0.7.3",
    "jinja2>=3.1.6",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "apscheduler>=3.10.0",
]
